+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ``helpers.nearest_dict_value()`` function, to find key in a dictionary that's nearest to
 a given number and return the corresponding value.
 - ``loss_functions.standard_normal_loss_dict()`` function, to build a dictionary of loss-function values.
+- Support for negative binomial demand distributions in ``DemandSource``.
 
 ### Fixed
 - Bug in ``helpers.ensure_list_for_time_period()`` that caused it to handle numpy arrays improperly.